requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.12.5
python-dotenv==1.0.0
qdrant-client==1.12.0
cohere==5.5.8
openai>=1.12.0
requests==2.31.0
beautifulsoup4==4.12.2
tiktoken==0.7.0
langchain==0.3.0
slowapi==0.1.9
limits==5.6.0